requirements.txt
# 核心科学计算库
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0

# 有限元和高性能计算
petsc4py>=3.17.0
mpi4py>=3.1.0
metis>=0.2.0

# 机器学习和优化
scikit-learn>=1.0.0
torch>=1.10.0
tensorflow>=2.8.0

# GPU加速
cupy-cuda11x  # 根据CUDA版本选择，可选: cupy-cuda11x, cupy-cuda12x
pyopencl>=2022.1.0
torchvision>=0.11.0
torchaudio>=0.11.0

# ML优化专用库
optuna>=3.0.0          # 超参数优化
gpytorch>=1.9.0        # 高斯过程
lightgbm>=3.3.0        # 梯度提升
xgboost>=1.6.0         # XGBoost
catboost>=1.1.0        # CatBoost

# 深度学习和神经网络
keras>=2.8.0           # 高级神经网络API
transformers>=4.20.0   # 预训练模型
accelerate>=0.12.0     # 分布式训练

# 贝叶斯优化和不确定性量化
emukit>=0.4.0          # 贝叶斯优化
botorch>=0.6.0         # 贝叶斯优化（PyTorch）
pymc3>=3.11.0          # 贝叶斯建模

# 主动学习和数据管理
modAL>=0.4.0           # 主动学习
libact>=2.1.0          # 主动学习库
dask>=2022.1.0         # 并行计算框架

# 网格处理和可视化
meshio>=5.0.0
pyvista>=0.35.0
vtk>=9.0.0

# 并行计算和性能优化
numba>=0.56.0
cython>=0.29.0
joblib>=1.1.0

# 数值求解器
pypardiso>=0.3.0
pymumps>=0.2.0

# 符号计算
sympy>=1.10.0

# 数据分析和处理
pandas>=1.3.0
h5py>=3.6.0

# 测试和开发工具
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# 可选依赖（用于高级功能）
# 如果安装失败，可以注释掉这些行
# hypre>=2.20.0  # 代数多重网格
# superlu>=5.2.0  # 稀疏矩阵分解
# mumps>=5.2.0    # 并行稀疏求解器 